Make RPC call to abandon nested stack
Due to stacks being processable by multiple engines, resources in a
nested stack cannot be marked for retainment from the parent resource
of the nested stack. This patch makes the parent resource of the nested
stack call stack abandon instead of stack delete through the RPC client
so that nested stacks are deleted while nested stack resources are
retained.

Closes-bug: #1509912

+++ heat/engine/resources/stack_resource.py
@@ -54,6 +54,7 @@ class StackResource(resource.Resource):
 
     def __init__(self, name, json_snippet, stack):
         super(StackResource, self).__init__(name, json_snippet, stack)
+        self.nested_abandon_in_progress = False
         self._nested = None
         self.resource_info = None
 
@@ -455,7 +456,10 @@ class StackResource(resource.Resource):
         stack_identity = dict(stack.identifier())
 
         try:
-            self.rpc_client().delete_stack(self.context, stack_identity)
+            if self.nested_abandon_in_progress:
+                self.rpc_client().abandon_stack(self.context, stack_identity)
+            else:
+                self.rpc_client().delete_stack(self.context, stack_identity)
         except Exception as ex:
             self.rpc_client().ignore_error_named(ex, 'NotFound')
 
@@ -509,6 +513,7 @@ class StackResource(resource.Resource):
         return self._check_status_complete(self.CHECK)
 
     def prepare_abandon(self):
+        self.nested_abandon_in_progress = True
         nested_stack = self.nested()
         if nested_stack:
             return self.nested().prepare_abandon()
